Support Enhanced Device Request with lower-case PhotoID + NFC fixes.
If the mdoc supports Extended Request according to 18013-5 Second
Edition, use it to request either an mDL or a PhotoID (lower-case for
now).

Also fix NFC engagement for Android devices where a disambiguation
dialog is shown for the user to select which wallet to use as
explained in
[this commit](https://github.com/openwallet-foundation/multipaz/commit/6cf5e818148f6f77ce231e363ba3e9ef095a6cc6).

Bump Multipaz requirement to an unreleased version, we'll change it to
Multipaz 0.96.0 when that version is released early next month.

Test: Manually tested with Apple Wallet Digital ID, see https://youtu.be/cRYTFppuZRA
Signed-off-by: David Zeuthen <zeuthen@google.com>